+++ src/Phoenix.jl
@@ -19,7 +19,7 @@ module Phoenix
 # for the end-user.
 
 importall ScientificCameras
-import ScientificCameras: ScientificCamera, ROI
+import ScientificCameras: TimeoutError, ScientificCamera, ROI
 using ScientificCameras.PixelFormats
 export
     PHXError